Improve manual test abstractions
